+++ src/com/eatery/models/ChocoChipDecorator.java
@@ -6,7 +6,7 @@ public class ChocoChipDecorator extends AddOnsDecorator{
     }
 
     @Override
-    public int getTotalPrice() {
+    public double getTotalPrice() {
         return super.getTotalPrice() + 5;
     }
 }
