+++ src/com/eatery/models/DrinksBuilder.java
@@ -3,11 +3,11 @@ package com.eatery.models;
 public class DrinksBuilder extends ItemBuilder{
     private String name;
     private boolean isVisible = false;
-    private int price;
+    private double price;
     private int id = -1;
     private int offers = 0;
 
-    public DrinksBuilder(String name, int price){
+    public DrinksBuilder(String name, double price){
         this.name = name;
         this.price = price;
     }
