+++ src/com/eatery/models/Fries.java
@@ -2,12 +2,12 @@ package com.eatery.models;
 
 public class Fries extends Item {
 
-    Fries(int itemId, String itemName, int price, Boolean visibility, int offers) {
+    Fries(int itemId, String itemName, double price, Boolean visibility, int offers) {
         super(itemId, itemName, price, visibility, offers, "fries");
     }
 
     @Override
-    public int getTotalPrice() {
+    public double getTotalPrice() {
         return 0;
     }
