+++ src/com/eatery/models/MintDecorator.java
@@ -6,7 +6,7 @@ public class MintDecorator extends AddOnsDecorator {
 
     }
         @Override
-        public int getTotalPrice() {
+        public double getTotalPrice() {
             return super.getTotalPrice() + 6;
         }
